+++ TR5Main/Objects/TR3/Vehicles/kayak_info.h
@@ -3,7 +3,7 @@
 
 struct KayakInfo 
 {
-	int TurnRate;
+	float TurnRate;
 
 	int Velocity;
 	int FrontVerticalVelocity;
